requirements.txt
fastapi
pydantic
uvicorn
watchfiles
psycopg2
pyyaml
psutil
jinja2
python-multipart
redis